+++ src/agent/onefuzz-task/src/local/example_templates/libfuzzer_basic.yml
@@ -26,7 +26,7 @@ tasks:
     crashes: *crash
     reports: "./reports"
     unique_reports: "./unique_reports"
-    no_repro: "./noe_repro"
+    no_repro: "./no_repro"
     check_fuzzer_help: true
 
   - type: "Coverage"
@@ -36,6 +36,3 @@ tasks:
     input_queue: *inputs
     readonly_inputs: [*inputs]
     coverage: "./coverage"
-
-  # - type: Analysis
-  #   <<: *target_args
