+++ src/agent/onefuzz-task/src/local/libfuzzer_merge.rs
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-use std::path::PathBuf;
+use std::{collections::HashMap, path::PathBuf};
 
 use crate::{
     local::common::{
@@ -13,12 +13,19 @@ use crate::{
     tasks::{
         config::CommonConfig,
         merge::libfuzzer_merge::{spawn, Config},
+        utils::default_bool_true,
     },
 };
 use anyhow::Result;
+use async_trait::async_trait;
 use clap::{Arg, ArgAction, Command};
 use flume::Sender;
+use futures::future::OptionFuture;
+use onefuzz::syncdir::SyncedDir;
+use schemars::JsonSchema;
 use storage_queue::QueueClient;
+
+use super::template::{RunContext, Template};
 
 pub fn build_merge_config(
     args: &clap::ArgMatches,
@@ -85,4 +92,63 @@ pub fn args(name: &'static str) -> Command {
     Command::new(name)
         .about("execute a local-only libfuzzer crash report task")
         .args(&build_shared_args())
+}
+
+#[derive(Debug, Serialize, Deserialize, Clone, JsonSchema)]
+pub struct LibfuzzerMerge {
+    target_exe: PathBuf,
+    target_env: HashMap<String, String>,
+    target_options: Vec<String>,
+    input_queue: Option<PathBuf>,
+    inputs: Vec<PathBuf>,
+    unique_inputs: PathBuf,
+    preserve_existing_outputs: bool,
+
+    #[serde(default = "default_bool_true")]
+    check_fuzzer_help: bool,
+}
+
+#[async_trait]
+impl Template for LibfuzzerMerge {
+    async fn run(&self, context: &RunContext) -> Result<()> {
+        let input_q_fut: OptionFuture<_> = self
+            .input_queue
+            .iter()
+            .map(|w| context.monitor_dir(w))
+            .next()
+            .into();
+        let input_q = input_q_fut.await.transpose()?;
+
+        let libfuzzer_merge = crate::tasks::merge::libfuzzer_merge::Config {
+            target_exe: self.target_exe.clone(),
+            target_env: self.target_env.clone(),
+            target_options: self.target_options.clone(),
+            input_queue: input_q,
+            inputs: self
+                .inputs
+                .iter()
+                .enumerate()
+                .map(|(index, roi_pb)| {
+                    context.to_monitored_sync_dir(format!("inputs_{index}"), roi_pb)
+                })
+                .collect::<Result<Vec<SyncedDir>>>()?,
+            unique_inputs: context
+                .to_monitored_sync_dir("unique_inputs", self.unique_inputs.clone())?,
+            preserve_existing_outputs: self.preserve_existing_outputs,
+
+            check_fuzzer_help: self.check_fuzzer_help,
+
+            common: CommonConfig {
+                task_id: uuid::Uuid::new_v4(),
+                ..context.common.clone()
+            },
+        };
+
+        context
+            .spawn(
+                async move { crate::tasks::merge::libfuzzer_merge::spawn(libfuzzer_merge).await },
+            )
+            .await;
+        Ok(())
+    }
 }
